+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this package are documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.6.0] - UNRELEASED
+
+### Added
+
+- Added `ControllerType` property to `Glyph` to inform components about the device the glyph intends to represent.
+	- All GUID and Template maps represent `Joystick` glyphs, Hardware maps are usually `Joystick` except for `Keyboard`, `Mouse`, and `Unknown` hardware definition.
+
 ## [1.5.0] - 2024-05-30
 
 ### Added